requirements.txt
pyttsx3==2.90
SpeechRecognition==3.10.0
plyer==2.1.0
# youtube-search-python==1.6.2  # Optional - has compatibility issues, using fallback method
wikipedia==1.4.0
pyautogui==0.9.54
pyjokes==0.6.0
speedtest-cli==2.1.3
python-dotenv==1.0.0
newsapi-python==0.2.7
requests==2.31.0
wolframalpha==5.0.0
# PyAudio==0.2.11  # Commented out due to Windows build issues - install separately if needed

# New features dependencies
opencv-python==4.8.1.78
sounddevice==0.4.6
soundfile==0.12.1
yt-dlp==2023.12.30
googletrans==4.0.0rc1
Pillow==10.0.0